requirements.txt
Django==5.2.4
djangorestframework
celery
redis
libvirt-python
PyYAML
djangorestframework-simplejwt
